requirements.txt
aioredis==1.3.1
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.2.1
Automat==20.2.0
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
colorama==0.4.3
configparser==4.0.2
constantly==15.1.0
crayons==0.3.0
cryptography==2.8
daphne==2.4.1
defusedxml==0.6.0
dj-rest-auth==0.1.4
Django==3.0.3
django-allauth==0.41.0
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
Markdown==3.2.1
msgpack==0.6.2
oauthlib==3.1.0
pkg-resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==2.0.0
pyOpenSSL==19.1.0
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==20.1.1
urllib3==1.25.8
webdriver-manager==2.3.0
zope.interface==4.7.1
